+++ python/ray/includes/common.pxd
@@ -12,7 +12,7 @@ from ray.includes.unique_ids cimport (
 )
 
 
-cdef extern from "ray/status.h" namespace "ray" nogil:
+cdef extern from "ray/common/status.h" namespace "ray" nogil:
     cdef cppclass StatusCode:
         pass
 
@@ -68,7 +68,7 @@ cdef extern from "ray/status.h" namespace "ray" nogil:
     cdef CRayStatus RayStatus_Invalid "Status::Invalid"()
 
 
-cdef extern from "ray/status.h" namespace "ray::StatusCode" nogil:
+cdef extern from "ray/common/status.h" namespace "ray::StatusCode" nogil:
     cdef StatusCode StatusCode_OK "OK"
     cdef StatusCode StatusCode_OutOfMemory "OutOfMemory"
     cdef StatusCode StatusCode_KeyError "KeyError"
