+++ src/ray/core_worker/object_interface.h
@@ -5,7 +5,7 @@
 #include "plasma/client.h"
 #include "ray/common/buffer.h"
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 
 namespace ray {
