+++ src/ray/core_worker/task_execution.h
@@ -3,7 +3,7 @@
 
 #include "common.h"
 #include "ray/common/buffer.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 
 namespace ray {
