+++ src/ray/core_worker/task_interface.h
@@ -4,7 +4,7 @@
 #include "common.h"
 #include "ray/common/buffer.h"
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 
 namespace ray {
