+++ src/ray/gcs/client.h
@@ -7,7 +7,7 @@
 #include "ray/gcs/asio.h"
 #include "ray/gcs/tables.h"
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 #include "ray/util/logging.h"
 
 namespace ray {
