+++ src/ray/gcs/redis_context.h
@@ -6,7 +6,7 @@
 #include <unordered_map>
 
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 #include "ray/util/logging.h"
 
 #include "ray/gcs/format/gcs_generated.h"
