+++ src/ray/gcs/redis_module/ray_redis_module.cc
@@ -4,7 +4,7 @@
 #include "ray/common/common_protocol.h"
 #include "ray/gcs/format/gcs_generated.h"
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 #include "ray/util/logging.h"
 #include "redis_string.h"
 #include "redismodule.h"
