+++ src/ray/object_manager/object_buffer_pool.h
@@ -13,7 +13,7 @@
 #include "plasma/client.h"
 
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 
 namespace ray {
