+++ src/ray/object_manager/object_manager.h
@@ -17,7 +17,7 @@
 
 #include "ray/common/client_connection.h"
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 
 #include "ray/object_manager/connection_pool.h"
 #include "ray/object_manager/format/object_manager_generated.h"
