+++ src/ray/object_manager/object_store_notification_manager.h
@@ -12,7 +12,7 @@
 #include "plasma/client.h"
 
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 
 #include "ray/object_manager/object_directory.h"
