+++ src/ray/object_manager/test/object_manager_test.cc
@@ -3,7 +3,7 @@
 
 #include "gtest/gtest.h"
 
-#include "ray/status.h"
+#include "ray/common/status.h"
 
 #include "ray/object_manager/object_manager.h"
