+++ src/ray/raylet/lineage_cache.h
@@ -9,7 +9,7 @@
 #include "ray/raylet/task.h"
 #include "ray/gcs/tables.h"
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 // clang-format on
 
 namespace ray {
