+++ src/ray/raylet/mock_gcs_client.h
@@ -13,7 +13,7 @@
 #include <boost/bind.hpp>
 
 #include "ray/common/id.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 
 namespace ray {
