+++ src/ray/raylet/worker_pool.cc
@@ -7,7 +7,7 @@
 
 #include "ray/ray_config.h"
 #include "ray/stats/stats.h"
-#include "ray/status.h"
+#include "ray/common/status.h"
 #include "ray/util/logging.h"
 
 namespace {
